+++ src/components/LobbyComponent.vue
@@ -4,7 +4,7 @@ import type { Ref } from 'vue';
 import { useRouter } from 'vue-router';
 import { LobbyClient } from 'boardgame.io/client';
 import Button from 'primevue/button';
-import { useUser } from '@clerk/vue';
+import { useClerkUser } from '@/composables/useClerkUser';
 
 import LobbyMatch from '@/components/LobbyMatch.vue';
 import type { GameSetupData } from '@/game/Game';
@@ -20,7 +20,7 @@ import { type LobbyAPI } from 'boardgame.io';
 const matches: Ref<EnhancedMatch[]> = ref([]);
 const lastFetched = ref();
 const server = getServerURL();
-const { user } = useUser();
+const { clerkUsername } = useClerkUser();
 
 const dateFormatter = new Intl.DateTimeFormat('en-US', {
   month: 'short',
@@ -80,7 +80,7 @@ const shouldHighlight = (match: EnhancedMatch) => {
   let yourTurn;
   if (activePlayers) {
     const playerIndex = Object.values(players).findIndex(
-      (player) => player.name === user.value?.username,
+      (player) => player.name === clerkUsername.value,
     );
     yourTurn = activePlayers[playerIndex] === 'planning';
   }
@@ -92,14 +92,14 @@ const yourMatches: Ref<EnhancedMatch[]> = ref([]);
 const openMatches: Ref<EnhancedMatch[]> = ref([]);
 const remainingMatches: Ref<EnhancedMatch[]> = ref([]);
 
-watch([matches, user], () => {
+watch([matches, clerkUsername], () => {
   const newYourMatches: EnhancedMatch[] = [];
   const newOpenMatches: EnhancedMatch[] = [];
   const newRemainingMatches: EnhancedMatch[] = [];
 
   matches.value.forEach((match) => {
     if (
-      match.players.some((p) => p.name && p.name === user.value?.username) &&
+      match.players.some((p) => p.name && p.name === clerkUsername.value) &&
       newYourMatches.length < 6
     ) {
       newYourMatches.push(match);
